+++ types/src/lib.rs
@@ -6,7 +6,7 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 #![cfg_attr(doc_cfg, feature(doc_cfg))]
 #![deny(missing_docs, warnings)]
-#![allow(clippy::redundant_pub_crate)]
+#![allow(clippy::redundant_pub_crate, clippy::module_name_repetitions)]
 
 extern crate alloc;
 #[cfg(feature = "std")]