+++ client/CHANGELOG.md
@@ -19,6 +19,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
+## 2.0.1-rc.7 - 2023-XX-XX
+
+### Changed
+
+- `DatabaseProvider` renamed to `StorageProvider`;
+- `StrongholdDatabaseProvider` renamed to `StrongholdStorageProvider`;
+- `high_level` module moved to `api`;
+
 ## 2.0.1-rc.6 - 2023-02-08
 
 ### Added